1.0.0.b1 WIP on change to JSON datafile + adding 'next action' feature
