+++ jgclark.Reviews/src/index.js
@@ -3,7 +3,7 @@
 //-----------------------------------------------------------------------------
 // Index for Reviews plugin
 // by Jonathan Clark
-// Last updated 2024-08-30 for v0.14.1, @jgclark
+// Last updated 2024-09-27 for v1.0.0.b1, @jgclark
 //-----------------------------------------------------------------------------
 
 // allow changes in plugin.json to trigger recompilation
@@ -11,28 +11,37 @@
 import pluginJson from '../plugin.json'
 import { getReviewSettings, type ReviewConfig } from './reviewHelpers'
 import {
-  makeFullReviewList,
   renderProjectLists
 } from './reviews'
+import {
+  generateAllProjectsList,
+  // makeFullReviewList,
+} from './reviewListHelpers'
 import { pluginUpdated, updateSettingData } from '@helpers/NPConfiguration'
 import { JSP, logDebug, logError, logInfo } from '@helpers/dev'
 import { editSettings } from '@helpers/NPSettings'
 import { isHTMLWindowOpen } from '@helpers/NPWindows'
 
 export {
-  logFullReviewList,
-  makeFullReviewList,
-  startReviews,
-  // nextReview,
   finishReview,
   finishReviewAndStartNextReview,
-  skipReview,
-  makeProjectLists,
+  makeProjectLists, // TODO: rename to displayProjectLists
+  // nextReview,
   redisplayProjectListHTML,
   renderProjectLists,
+  skipReview,
+  startReviews,
   toggleDisplayFinished,
   toggleDisplayOnlyDue
 } from './reviews'
+export {
+  generateAllProjectsList,
+  getNextNoteToReview, //  TODO: remove in time
+  getNextProjectsToReview, //  TODO: remove in time
+  logAllProjectsList,
+  logFullReviewList,
+  makeFullReviewList,
+} from './reviewListHelpers'
 export { removeAllDueDates } from '@helpers/NPParagraph'
 export {
   addProgressUpdate,
@@ -43,11 +52,6 @@ export {
 export {
   generateCSSFromTheme
 } from '@helpers/NPThemeToCSS'
-// export {
-//   setHTMLWinHeight,
-// } from '@helpers/NPWindows'
-
-// An earlier version had some functions in projectLists.js, but they had to be moved because of dependency issues.
 
 // Note: There are other possible exports, including:
 export { testFonts } from '../experiments/fontTests.js'
@@ -58,8 +62,7 @@ export {
   testCSSCircle,
   testRedToGreenInterpolation,
 } from './HTMLtests'
-export { onMessageFromHTMLView } from './pluginToHTMLBridge'
-
+export { onMessageFromHTMLView } from './pluginToHTMLBridge' // TODO: is this needed?
 const pluginID = 'jgclark.Reviews'
 
 export function init(): void {
@@ -85,7 +88,10 @@ export async function onSettingsUpdated(): Promise<void> {
   // Update the full - review - list in case there's a change in a relevant setting
   logDebug(pluginID, 'Have updated settings, so will recalc the review list and display...')
   const config: ReviewConfig = await getReviewSettings()
-  await makeFullReviewList(config, true)
+
+  // await makeFullReviewList(config, true)
+  await generateAllProjectsList(config, true)
+
   // If v3.11+, can now refresh Dashboard
   if (NotePlan.environment.buildVersion >= 1181) {
     if (isHTMLWindowOpen(pluginJson['plugin.id'])) {
