+++ jgclark.Reviews/src/reviewHelpers.js
@@ -2,43 +2,40 @@
 //-----------------------------------------------------------------------------
 // Helper functions for Review plugin
 // by Jonathan Clark
-// Last updated 2024-09-02 for v0.14.1, @jgclark
+// Last updated 2024-09-28 for v1.0.0.b1, @jgclark
 //-----------------------------------------------------------------------------
 
 //-----------------------------------------------------------------------------
 // Import Helper functions
-import moment from 'moment/min/moment-with-locales'
-import pluginJson from '../plugin.json'
-import { checkString } from '@helpers/checkType'
+// import pluginJson from '../plugin.json'
+import { type Progress } from './projectClass'
 import {
   calcOffsetDate,
-  daysBetween,
+  // daysBetween,
   getDateFromUnhyphenatedDateString,
   getDateObjFromDateString,
   getJSDateStartOfToday,
-  includesScheduledFutureDate,
+  // includesScheduledFutureDate,
   RE_ISO_DATE, RE_YYYYMMDD_DATE,
-  todaysDateISOString,
+  // todaysDateISOString,
   toISODateString,
 } from '@helpers/dateTime'
-import { localeRelativeDateFromNumber } from '@helpers/NPdateTime'
 import { clo, JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
-import { getFolderFromFilename } from '@helpers/folders'
-import { createOpenOrDeleteNoteCallbackUrl, createRunPluginCallbackUrl, displayTitle, getContentFromBrackets, getStringFromList } from '@helpers/general'
 import {
-  makeSVGPercentRing,
-  redToGreenInterpolation,
-} from '@helpers/HTMLView'
+  createRunPluginCallbackUrl, displayTitle,
+} from '@helpers/general'
 import { noteHasFrontMatter, setFrontMatterVars } from '@helpers/NPFrontMatter'
-import { removeAllDueDates } from '@helpers/NPParagraph'
-import { findEndOfActivePartOfNote, findStartOfActivePartOfNote } from '@helpers/paragraph'
-import { encodeRFC3986URIComponent } from '@helpers/stringTransforms'
+// import { removeAllDueDates } from '@helpers/NPParagraph'
 import {
-  getInputTrimmed,
-  inputIntegerBounded,
+  findEndOfActivePartOfNote,
+  // findStartOfActivePartOfNote
+} from '@helpers/paragraph'
+import {
+// getInputTrimmed,
+// inputIntegerBounded,
   showMessage
 } from '@helpers/userInput'
-import { isDone, isOpen } from '@helpers/utils'
+import { isOpen } from '@helpers/utils'
 
 //------------------------------
 // Config setup
@@ -71,6 +68,8 @@ export type ReviewConfig = {
   height: number,
   archiveUsingFolderStructure: boolean,
   removeDueDatesOnPause: boolean,
+  nextActionTag: string,
+  displayNextActions: boolean,
   _logLevel: string,
   _logTimer: boolean,
   _logLevel: string,
@@ -82,8 +81,8 @@ export type ReviewConfig = {
  * @return {ReviewConfig} object with configuration
  */
 export async function getReviewSettings(): Promise<ReviewConfig> {
-  // logDebug(pluginJson, `Start of getReviewSettings()`)
   try {
+    // logDebug('getReviewSettings', `Starting ...`)
     // Get settings
     const config: ReviewConfig = await DataStore.loadJSON('../jgclark.Reviews/settings.json')
 
@@ -214,19 +213,12 @@ export function getFieldParagraphsFromNote(note: TNote, fieldName: string): Arra
   return matchArr
 }
 
-export type Progress = {
-  lineIndex: number,
-  percentComplete: number,
-  date: Date,
-  comment: string
-}
-
 /**
  * Return the (paragraph index of) the most recent progress line from the array, based upon the most recent YYYYMMDD or YYYY-MM-DD date found. If it can't find any it default to the first paragraph.
  * @param {Array<TParagraph>} progressParas
  * @returns {number} lineIndex of the most recent line
  */
-function processMostRecentProgressParagraph(progressParas: Array<TParagraph>): Progress {
+export function processMostRecentProgressParagraph(progressParas: Array<TParagraph>): Progress {
   try {
     let lastDate = new Date('0000-01-01') // earliest possible YYYY-MM-DD date
     // let lastIndex = 0 // Default to returning first line
@@ -293,13 +285,13 @@ function processMostRecentProgressParagraph(progressParas: Array<TParagraph>): P
  * Works out which line (if any) of the current note is project-style metadata line, defined as
  * - line starting 'project:' or 'medadata:'
  * - first line containing a @review() or @reviewed() mention
- * - first line starting with a hashtag
+ * - first line starting with a hashtag.
  * If these can't be found, then create a new line after the title, or in the 'metadata:' field if present in the frontmatter.
  * @author @jgclark
  *
  * @param {TNote} note to use
  * @param {string} metadataLinePlaceholder optional to use if we need to make a new metadata line
- * @returns {number} the line number for the metadata line
+ * @returns {number} the line number for the existing or new metadata line
  */
 export function getOrMakeMetadataLine(note: TNote, metadataLinePlaceholder: string = ''): number {
   try {
@@ -352,897 +344,22 @@ export function getOrMakeMetadataLine(note: TNote, metadataLinePlaceholder: stri
   }
 }
 
-
-//-----------------------------------------------------------------------------
-
-/**
- * Define 'Project' class to use in GTD.
- * Holds title, last reviewed date, due date, review interval, completion date,
- * number of closed, open & waiting for tasks.
- *
- * @example To create a project instance for a note call 'const x = new Project(note)'
- * @author @jgclark
- */
-export class Project {
-  // Types for the class instance properties
-  note: TNote
-  filename: string
-  metadataPara: TParagraph
-  noteType: string // #project, #area, etc.
-  title: string
-  startDate: ?Date
-  dueDate: ?Date
-  dueDays: number = NaN
-  reviewedDate: ?Date
-  reviewInterval: string // later will default to '1w' if needed
-  nextReviewDate: ?Date
-  nextReviewDateStr: ?string // can be set by user (temporarily) but not otherwise populated
-  nextReviewDays: number = NaN
-  completedDate: ?Date
-  completedDuration: ?string // string description of time to completion, or how long ago completed
-  cancelledDate: ?Date
-  cancelledDuration: ?string // string description of time to cancellation, or how long ago cancelled
-  // finishedDays: number = NaN // days until project was completed or cancelled
-  isCompleted: boolean = false
-  openTasks: number
-  completedTasks: number
-  waitingTasks: number
-  futureTasks: number
-  isCancelled: boolean = false
-  isPaused: boolean = false
-  folder: string
-  percentComplete: number = NaN
-  lastProgressComment: string = '' // e.g. "Progress: 60@20220809: comment
-  mostRecentProgressLineIndex: number = NaN
-  ID: string // required when making HTML views
-
-  constructor(note: TNote, noteTypeTag: string = '', checkEditor: boolean = true) {
-    try {
-      if (note == null || note.title == null) {
-        throw new Error('Error in constructor: invalid note passed')
-      }
-      this.title = note.title
-      this.filename = note.filename
-      logDebug('Project constructor', `Starting for Note: ${this.filename} type ${noteTypeTag}:`)
-      this.folder = getFolderFromFilename(note.filename)
-
-      // Make a (nearly) unique number for this instance (needed for the addressing the SVG circles) -- I can't think of a way of doing this neatly to create one-up numbers, that doesn't create clashes when re-running over a subset of notes
-      this.ID = String(Math.round((Math.random()) * 99999))
-
-      // Sometimes we're called just after a note has been updated in the Editor. So check to see if note is open in Editor, and if so use that version, which could be newer.
-      // (Unless 'checkEditor' false, to avoid triggering 'You are running this on an async thread' warnings.)
-      let paras: Array<TParagraph>
-      if (checkEditor && Editor && Editor.note && (Editor.note.filename === note.filename)) {
-        const editorNote: CoreNoteFields = Editor.note
-        paras = editorNote.paragraphs
-        this.note = Editor.note // Note: not plain Editor, as otherwise it isn't the right type and will throw app run-time errors later.
-        const timeSinceLastEdit: number = Date.now() - editorNote.versions[0].date
-        logDebug('Project constructor', `- using EDITOR for (${Editor.filename}), last updated ${String(timeSinceLastEdit)}ms ago.} `)
-      } else {
-        // read note from DataStore in the usual way
-        paras = note.paragraphs
-        this.note = note
-        // logDebug('Project constructor', `- read note from datastore `)
-      }
-
-      const metadataLineIndex = getOrMakeMetadataLine(note)
-      this.metadataPara = paras[metadataLineIndex]
-      let mentions: $ReadOnlyArray<string> = note.mentions ?? [] // Note: can be out of date, and I can't find a way of fixing this, even with updateCache()
-      let hashtags: $ReadOnlyArray<string> = note.hashtags ?? [] // Note: can be out of date
-      const metadataLine = paras[metadataLineIndex].content
-      if (mentions.length === 0) {
-        logDebug('Project constructor', `- Grr: .mentions empty: will use metadata line instead`)
-        // Note: If necessary, fall back to getting mentions just from the metadataline
-        mentions = (`${metadataLine} `).split(' ').filter((f) => f[0] === '@')
-      }
-      if (hashtags.length === 0) {
-        hashtags = (`${metadataLine} `).split(' ').filter((f) => f[0] === '#')
-      }
-
-      // work out noteType:
-      // - if noteTypeTag given, then use that
-      // - else first or second hashtag in note
-      try {
-        this.noteType = (noteTypeTag)
-          ? noteTypeTag
-          : (hashtags[0] !== '#paused')
-            ? hashtags[0]
-            : (hashtags[1])
-              ? hashtags[1]
-              : ''
-      } catch (e) {
-        this.noteType = ''
-        logWarn('Project constructor', `- found no noteType for '${this.title}' in folder ${this.folder}`)
-      }
-
-      // read in various metadata fields (if present)
-      // FIXME: doesn't pick up reviewed() if not in metadata line
-      let tempStr = getParamMentionFromList(mentions, checkString(DataStore.preference('startMentionStr')))
-      this.startDate = tempStr !== '' ? getDateObjFromDateString(tempStr) : undefined
-      // read in due date (if found)
-      tempStr = getParamMentionFromList(mentions, checkString(DataStore.preference('dueMentionStr')))
-      this.dueDate = tempStr !== '' ? getDateObjFromDateString(tempStr) : undefined
-      // read in reviewed date (if found)
-      tempStr = getParamMentionFromList(mentions, checkString(DataStore.preference('reviewedMentionStr')))
-      this.reviewedDate = tempStr !== '' ? getDateObjFromDateString(tempStr) : undefined
-      // read in completed date (if found)
-      tempStr = getParamMentionFromList(mentions, checkString(DataStore.preference('completedMentionStr')))
-      this.completedDate = tempStr !== '' ? getDateObjFromDateString(tempStr) : undefined
-      // read in cancelled date (if found)
-      tempStr = getParamMentionFromList(mentions, checkString(DataStore.preference('cancelledMentionStr')))
-      this.cancelledDate = tempStr !== '' ? getDateObjFromDateString(tempStr) : undefined
-      // read in review interval (if found)
-      const tempIntervalStr = getParamMentionFromList(mentions, checkString(DataStore.preference('reviewIntervalMentionStr')))
-      // $FlowIgnore[incompatible-type]
-      this.reviewInterval = tempIntervalStr !== '' ? getContentFromBrackets(tempIntervalStr) : '1w'
-      // read in nextReview date (if found)
-      tempStr = getParamMentionFromList(mentions, checkString(DataStore.preference('nextReviewMentionStr')))
-      if (tempStr !== '') {
-        this.nextReviewDate = getDateObjFromDateString(tempStr)
-        if (this.nextReviewDate) {
-          this.nextReviewDateStr = toISODateString(this.nextReviewDate)
-          logDebug('Project constructor', `- found '@nextReview(${this.nextReviewDateStr})' = ${String(this.nextReviewDate)}`)
-        } else {
-          logWarn('Project constructor', `- couldn't get valid date from  '@nextReview(${tempStr})'`)
-        }
-      }
-
-      // count tasks (includes both tasks and checklists)
-      // Note: excludes future tasks
-      this.openTasks = paras.filter(isOpen).length
-      this.completedTasks = paras.filter(isDone).length
-      this.waitingTasks = paras.filter(isOpen).filter((p) => p.content.match('#waiting')).length
-      this.futureTasks = paras.filter(isOpen).filter((p) => includesScheduledFutureDate(p.content)).length
-
-      // make project completed if @completed(date) set
-      if (this.completedDate != null) {
-        this.isCompleted = true
-        this.nextReviewDays = NaN
-      }
-      // make project cancelled if @cancelled(date) set
-      if (this.cancelledDate != null) {
-        this.isCancelled = true
-        this.nextReviewDays = NaN
-      }
-      // make project paused if #paused
-      if (getStringFromList(hashtags, '#paused') !== '') {
-        this.isPaused = true
-        this.nextReviewDays = NaN
-      }
-
-      // calculate the durations from these dates
-      this.calcDurations()
-      // if not finished, calculate next review dates
-      if (!this.isCancelled && !this.isCompleted) {
-        this.calcNextReviewDate()
-      }
-
-      // Find progress field lines (if any) and process
-      this.processProgressLines()
-
-      // If percentComplete not set via progress line, then calculate
-      const totalTasks = this.completedTasks + this.openTasks - this.futureTasks
-      if (totalTasks > 0) {
-        // use 'floor' not 'round' to ensure we don't get to 100% unless really everything is done
-        this.percentComplete = Math.floor((this.completedTasks / totalTasks) * 100)
-      } else {
-        this.percentComplete = NaN
-      }
-
-      // logDebug('Project constructor', `project(${this.title}) -> ID ${this.ID} / ${this.nextReviewDateStr ?? '-'} / ${String(this.nextReviewDays)} / ${this.isCompleted ? ' completed' : ''}${this.isCancelled ? ' cancelled' : ''}${this.isPaused ? ' paused' : ''}`)
-
-      if (this.title.includes('(TEST)')) {
-        logDebug('Project constructor', `- for '${this.title}' (${this.filename}) in folder ${this.folder}`)
-        logDebug('Project constructor', `  - metadataLine = ${metadataLine}`)
-        logDebug('Project constructor', `  - mentions: ${String(mentions)}`)
-        // logDebug('Project constructor', `  - altMentions: ${String(altMentions)}`)
-        logDebug('Project constructor', `  - hashtags: ${String(hashtags)}`)
-        // logDebug('Project constructor', `  - altHashtags: ${String(altHashtags)}`)
-        logDebug('Project constructor', `  - open: ${String(this.openTasks)}`)
-        logDebug('Project constructor', `  - completed: ${String(this.completedTasks)}`)
-        logDebug('Project constructor', `  - % complete = ${String(this.percentComplete)}`)
-      }
-    }
-    catch (error) {
-      logError('Project constructor', error.message)
-    }
-  }
-
-  /**
-   * Is this project ready for review?
-   * Return true if review is due and not archived or completed
-   * @return {boolean}
-   */
-  get isReadyForReview(): boolean {
-    // logDebug(pluginJson, `isReadyForReview: ${this.title}:  ${String(this.nextReviewDays)} ${String(this.isPaused)}`)
-    // $FlowFixMe[invalid-compare]
-    return !this.isPaused && this.nextReviewDays != null && !isNaN(this.nextReviewDays) && this.nextReviewDays <= 0
-  }
-
-  /**
-   * From the project metadata read in, calculate due/finished durations
-   */
-  calcDurations(): void {
-    try {
-      const now = new moment().toDate() // use moment instead of `new Date` to ensure we get a date in the local timezone
-      // Calculate # days until due
-      this.dueDays =
-        this.dueDate != null
-        ? daysBetween(now, this.dueDate)
-        : NaN
-
-      // Calculate durations or time since cancel/complete
-      // logDebug('calcDurations', String(this.startDate ?? 'no startDate'))
-      if (this.startDate) {
-        const momTSD = moment(this.startDate)
-        if (this.completedDate != null) {
-          this.completedDuration = `after ${momTSD.to(moment(this.completedDate), true)}`
-          // logDebug('calcDurations', `-> completedDuration = ${this.completedDuration}`)
-        }
-        else if (this.cancelledDate != null) {
-          this.cancelledDuration = `after ${momTSD.to(moment(this.cancelledDate), true)}`
-          // logDebug('calcDurations', `-> cancelledDuration = ${this.cancelledDuration}`)
-        }
-      }
-      else {
-        if (this.completedDate != null) {
-          this.completedDuration = moment(this.completedDate).fromNow() // ...ago
-          if (this.completedDuration.includes('hours')) {
-            this.completedDuration = 'today' // edge case
-          }
-          // logDebug('calcDurations', `-> completedDuration = ${this.completedDuration}`)
-        }
-        else if (this.cancelledDate != null) {
-          this.cancelledDuration = moment(this.cancelledDate).fromNow() // ...ago
-          if (this.cancelledDuration.includes('hours')) {
-            this.cancelledDuration = 'today' // edge case
-          }
-          // logDebug('calcDurations', `-> completedDuration = ${this.cancelledDuration}`)
-        }
-        else {
-          // Nothing to do
-          // logDebug('calcDurations', `No completed or cancelled dates.`)
-        }
-      }
-    } catch (error) {
-      logError('calcDurations', error.message)
-    }
-  }
-
-  calcNextReviewDate(): void {
-    try {
-      // Calculate next review due date, if there isn't already a nextReviewDate, and there's a review interval.
-      const now = new moment().toDate() // use moment instead of  `new Date` to ensure we get a date in the local timezone
-
-      // First check to see if project start is in future: if so set nextReviewDate to project start
-      if (this.startDate) {
-        const momTSD = moment(this.startDate)
-        if (momTSD.isAfter(now)) {
-          this.nextReviewDate = this.startDate
-          this.nextReviewDays = daysBetween(now, momTSD.toDate())
-          logDebug('calcNextReviewDate', `project start is in future (${momTSD.format('YYYY-MM-DD')}) -> ${String(this.nextReviewDays)} interval`)
-          return
-        }
-      }
-
-      // Now check to see if we have a specific nextReviewDate
-      if (this.nextReviewDateStr != null) {
-        this.nextReviewDays = daysBetween(now, this.nextReviewDateStr)
-        logDebug('calcNextReviewDate', `already had a nextReviewDateStr ${this.nextReviewDateStr ?? '?'} -> ${String(this.nextReviewDays)} interval`)
-      }
-      else if (this.reviewInterval != null) {
-        if (this.reviewedDate != null) {
-          this.nextReviewDate = calcNextReviewDate(this.reviewedDate, this.reviewInterval)
-          if (this.nextReviewDate != null) {
-            // this now uses moment and truncated (not rounded) date diffs in number of days
-            this.nextReviewDays = daysBetween(now, this.nextReviewDate)
-            // logDebug('calcNextReviewDate', `${String(this.reviewedDate)} + ${this.reviewInterval ?? ''} -> nextReviewDate: ${this.nextReviewDateStr ?? ''} = ${String(this.nextReviewDays) ?? '-'}`)
-          } else {
-            throw new Error(`nextReviewDate is null; reviewedDate = ${String(this.reviewedDate)}`)
-          }
-        } else {
-          // no next review date, so set at today
-          this.nextReviewDate = now
-          this.nextReviewDays = 0
-        }
-      }
-      // logDebug('calcNextReviewDate', `-> reviewedDate = ${String(this.reviewedDate)} / nextReviewDate = ${String(this.nextReviewDate)} / nextReviewDays = ${String(this.nextReviewDays)}`)
-    } catch (error) {
-      logError('calcNextReviewDate', error.message)
-    }
-  }
-
-  /**
-   * Prompt user for the details to make a progress line:
-   * - new % complete
-   * - new comment
-   * And add to the metadata area of the note
-   * @param {string} prompt message, to which is added the note title
-   */
-  async addProgressLine(prompt: string = 'Enter comment about current progress for'): Promise<void> {
-    try {
-      // Set insertion point for the new progress line to this paragraph,
-      // or if none exist, to the line after the current metadata line
-      let insertionIndex = this.mostRecentProgressLineIndex
-      if (isNaN(insertionIndex)) {
-        insertionIndex = findStartOfActivePartOfNote(this.note, true)
-        logDebug('Project::addProgressLine', `No progress paragraphs found, so will insert new progress line after metadata at line ${String(insertionIndex)}`)
-      } else {
-        // insertionIndex = processMostRecentProgressParagraph(getFieldParagraphsFromNote(this.note, 'progress'))
-        logDebug('Project::addProgressLine', `Will insert new progress line before most recent progress line at ${String(insertionIndex)}.`)
-      }
-
-      const message1 = `${prompt} '${this.title}'`
-      const resText = await getInputTrimmed(message1, 'OK', `Add Progress comment`)
-      if (!resText) {
-        logDebug('Project::addProgressLine', `No valid progress line given.`)
-        return
-      }
-      const comment = String(resText) // to keep flow happy
-
-      const message2 = (!isNaN(this.percentComplete)) ? `Enter project completion (as %; last was ${String(this.percentComplete)}%) if wanted` : `Enter project completion (as %) if wanted`
-      const resNum = await inputIntegerBounded('Add Progress % completion', message2, 100, 0)
-      let percentStr = ''
-      if (isNaN(resNum)) {
-        logDebug('Project::addProgressLine', `No percent completion given.`)
-      } else {
-        this.percentComplete = resNum
-        percentStr = String(resNum)
-      }
-
-      // Update the project's metadata
-      this.lastProgressComment = `${comment} (today)`
-      // logDebug('Project::addProgressLine', `-> line ${String(insertionIndex)}: ${this.percentComplete} / '${this.lastProgressComment}'`)
-      const newProgressLine = `Progress: ${percentStr}@${todaysDateISOString}: ${comment}`
-
-      // And write it to the Editor (if the note is open in it) ...
-      if (Editor && Editor.note && Editor.note.filename === this.note.filename) {
-        logDebug('Project::addProgressLine', `Writing '${newProgressLine}' to Editor at ${String(insertionIndex)}`)
-        Editor.insertParagraph(newProgressLine, insertionIndex, 'text')
-        // Also updateCache to make changes more quickly available elsewhere
-        await DataStore.updateCache(Editor, true)
-      }
-      // ... or the project's note
-      else {
-        logDebug('Project::addProgressLine', `Writing '${newProgressLine}' to project note '${this.note.filename}' at ${String(insertionIndex)}`)
-        this.note.insertParagraph(newProgressLine, insertionIndex, 'text')
-        // Also updateCache
-        await DataStore.updateCache(this.note, true)
-      }
-    } catch (error) {
-      logError(`Project::addProgressLine`, JSP(error))
-    }
-  }
-
-  /**
-   * Process the 'Progress:...' lines to retrieve metadata. Allowed forms are:
-   *   Progress: n@YYYYMMDD: progress messsage
-   *   Progress: n:YYYYMMDD: progress messsage
-   *   Progress: n:YYYY-MM-DD: progress messsage
-   *   Progress: n:YYYY-MM-DD: progress messsage
-   *   Progress: YYYYMMDD: progress messsage  [in which case % is calculated]
-   *   Progress: YYYY-MM-DD: progress messsage  [in which case % is calculated]
-   */
-  processProgressLines(): void {
-    // Get specific 'Progress' field lines
-    const progressParas = getFieldParagraphsFromNote(this.note, 'progress')
-
-    if (progressParas.length > 0) {
-      // Get the most recent progressItem from these lines
-      const progressItem: Progress = processMostRecentProgressParagraph(progressParas)
-      this.percentComplete = progressItem.percentComplete
-      this.lastProgressComment = progressItem.comment
-      this.mostRecentProgressLineIndex = progressItem.lineIndex
-      // logDebug('Project::processProgressLines', `  -> ${String(this.percentComplete)}% from progress line`)
-    } else {
-      logDebug('Project::processProgressLines', `- no progress fields found`)
-    }
-  }
-
-  /**
-   * Close a Project/Area note by updating the metadata and saving it:
-   * - adding @completed(<today's date>)
-   * @author @jgclark
-   * @returns {string} new machineSummaryLine or empty on failure
-   */
-  completeProject(): string {
-    try {
-      // update the metadata fields
-      // this.isActive = false
-      this.isCompleted = true
-      this.isCancelled = false
-      this.isPaused = false
-      this.completedDate = new moment().toDate() // use moment instead of `new Date` to ensure we get a date in the local timezone
-      this.calcDurations()
-
-      // re-write the note's metadata line
-      logDebug('completeProject', `Completing '${this.title}' ...`)
-      const newMetadataLine = this.generateMetadataLine()
-      logDebug('completeProject', `- metadata now '${newMetadataLine}'`)
-
-      // send update to Editor
-      // Note: Will need updating when supporting frontmatter for metadata
-      this.metadataPara.content = newMetadataLine
-      if (Editor && Editor.note && Editor.note === this.note) {
-        Editor.updateParagraph(this.metadataPara)
-        const res = DataStore.updateCache(this.note)
-      } else {
-        this.note.updateParagraph(this.metadataPara)
-        DataStore.updateCache(this.note, true)
-      }
-
-      const newMSL = this.machineSummaryLine()
-      logDebug('completeProject', `- returning mSL '${newMSL}'`)
-      return newMSL
-    }
-    catch (error) {
-      logError(pluginJson, `Error completing project for ${this.title}: ${error.message}`)
-      return ''
-    }
-  }
-
-  /**
-   * Cancel a Project/Area note by updating the metadata and saving it:
-   * - adding @cancelled(<today's date>)
-   * @author @jgclark
-   * @returns {string} new machineSummaryLine or empty on failure
-   */
-  cancelProject(): string {
-    try {
-      // update the metadata fields
-      // this.isActive = false
-      this.isCompleted = false
-      this.isCancelled = true
-      this.isPaused = false
-      this.cancelledDate = new moment().toDate()  // getJSDateStartOfToday() // use moment instead of `new Date` to ensure we get a date in the local timezone
-      this.calcDurations()
-
-      // re-write the note's metadata line
-      logDebug('cancelProject', `Cancelling '${this.title}' ...`)
-      const newMetadataLine = this.generateMetadataLine()
-      logDebug('cancelProject', `- metadata now '${newMetadataLine}'`)
-
-      // send update to Editor
-      // Note: Will need updating when supporting frontmatter for metadata
-      this.metadataPara.content = newMetadataLine
-      if (Editor && Editor.note && Editor.note === this.note) {
-        Editor.updateParagraph(this.metadataPara)
-        DataStore.updateCache(this.note, true)
-      } else {
-        this.note.updateParagraph(this.metadataPara)
-        DataStore.updateCache(this.note, true)
-      }
-
-      const newMSL = this.machineSummaryLine()
-      logDebug('cancelProject', `- returning mSL '${newMSL}'`)
-      return newMSL
-    }
-    catch (error) {
-      logError(pluginJson, `Error cancelling project for ${this.title}: ${error.message}`)
-      return ''
-    }
-  }
-
-  /**
-   * Cancel a Project/Area note by updating the metadata and saving it:
-   * - adding #paused
-   * @author @jgclark
-   * @returns {string} new machineSummaryLine or empty on failure
-   */
-  async togglePauseProject(): Promise<string> {
-    try {
-      // Get progress field details (if wanted)
-      await this.addProgressLine(this.isPaused ? 'Comment (if wanted) as you resume' : 'Comment (if wanted) as you pause')
-
-      // update the metadata fields
-      this.isCompleted = false
-      this.isCancelled = false
-      this.isPaused = !this.isPaused // toggle
-
-      // re-write the note's metadata line
-      logDebug('togglePauseProject', `Paused state now toggled to ${String(this.isPaused)} for '${this.title}' ...`)
-      const newMetadataLine = this.generateMetadataLine()
-      logDebug('togglePauseProject', `- metadata now '${newMetadataLine}'`)
-      // send update to Editor (if open)
-      // Note: Will need updating when supporting frontmatter for metadata
-      this.metadataPara.content = newMetadataLine
-      if (Editor && Editor.note && Editor.note === this.note) {
-        Editor.updateParagraph(this.metadataPara)
-        DataStore.updateCache(Editor.note, true)
-      } else {
-        this.note.updateParagraph(this.metadataPara)
-        DataStore.updateCache(this.note, true)
-      }
-
-      // if we want to remove all due dates on pause, then do that
-      if (this.isPaused) {
-        const config = await getReviewSettings()
-        if (config.removeDueDatesOnPause) {
-          logDebug('togglePauseProject', `- project now paused, and we want to remove due dates ...`)
-          const res = removeAllDueDates(this.filename)
-        }
-      }
-
-      const newMSL = this.machineSummaryLine()
-      logDebug('togglePauseProject', `- returning newMSL '${newMSL}'`)
-      return newMSL
-    }
-    catch (error) {
-      logError(pluginJson, `Error pausing project for ${this.title}: ${error.message}`)
-      return ''
-    }
-  }
-
-  /**
-   * Generate a one-line tab-sep summary line ready for Markdown note
-   */
-  generateMetadataLine(): string {
-    let output = this.noteType
-    output += ' '
-    output += this.isPaused ? '#paused ' : ''
-    // $FlowIgnore[incompatible-call]
-    output += this.startDate && this.startDate !== undefined ? `${checkString(DataStore.preference('startMentionStr'))}(${toISODateString(this.startDate)}) ` : ''
-    // $FlowIgnore[incompatible-call]
-    output += this.dueDate && this.startDate !== undefined ? `${checkString(DataStore.preference('dueMentionStr'))}(${toISODateString(this.dueDate)}) ` : ''
-    output +=
-      this.reviewInterval && this.reviewInterval !== undefined ? `${checkString(DataStore.preference('reviewIntervalMentionStr'))}(${checkString(this.reviewInterval)}) ` : ''
-    // $FlowIgnore[incompatible-call]
-    output += this.reviewedDate && this.reviewedDate !== undefined ? `${checkString(DataStore.preference('reviewedMentionStr'))}(${toISODateString(this.reviewedDate)}) ` : ''
-    // $FlowIgnore[incompatible-call]
-    output += this.completedDate && this.completedDate !== undefined ? `${checkString(DataStore.preference('completedMentionStr'))}(${toISODateString(this.completedDate)}) ` : ''
-    // $FlowIgnore[incompatible-call]
-    output += this.cancelledDate && this.cancelledDate !== undefined ? `${checkString(DataStore.preference('cancelledMentionStr'))}(${toISODateString(this.cancelledDate)}) ` : ''
-
-    return output
-  }
-
-  /**
-   * v2: Returns TSV line to go in full-review-list with just the data needed to filter output lists
-   * @return {string}
-   */
-  machineSummaryLine(): string {
-    try {
-      // next review in days
-      let output = (!this.isPaused && this.nextReviewDays != null && !isNaN(this.nextReviewDays)) ? String(this.nextReviewDays) : 'NaN'
-      output += '\t'
-      // due date in days
-      output += (!this.isPaused && this.dueDays != null && !isNaN(this.dueDays)) ? String(this.dueDays) : 'NaN'
-      // title
-      output += `\t${this.title}\t`
-      // folder
-      output += this.folder && this.folder !== undefined ? `${this.folder}\t` : '\t'
-      // note type, then other pseudo-tags
-      output += (this.noteType) ? `${this.noteType} ` : ''
-      output += this.isPaused ? '#paused' : ''
-      output += '\t'
-      output += (this.isCompleted)
-        ? 'finished'
-        : (this.isCancelled)
-          ? 'finished-cancelled'
-          : 'active'
-      return output
-    }
-    catch (error) {
-      logError('machineSummaryLine', error.message)
-      return '<error>' // for completeness
-    }
-  }
-
-  /**
-   * Returns title of note as folder name + link, also showing complete or cancelled where relevant.
-   * Supports 'Markdown' or 'HTML' styling or simpler 'list' styling
-   * @param {string} style 'Markdown' or 'HTML' or 'list'
-   * @param {boolean} includeFolderName whether to include folder name at the start of the entry.
-   * @return {string} - title as wikilink
-   */
-  decoratedProjectTitle(style: string, includeFolderName: boolean): string {
-    const folderNamePart = includeFolderName ? `${this.folder} / ` : ''
-    const titlePart = this.title ?? '(error, not available)'
-    // const titlePartEncoded = encodeURIComponent(this.title) ?? '(error, not available)'
-    switch (style) {
-      case 'Rich': {
-        // Method 1: make [[notelinks]] via x-callbacks
-        // Method 1a: x-callback using note title
-        // const noteOpenActionURL = createOpenOrDeleteNoteCallbackUrl(this.title, "title", "", "splitView", false)
-        // Method 1b: x-callback using filename
-        const noteOpenActionURL = createOpenOrDeleteNoteCallbackUrl(this.filename, "filename", "", null, false)
-        const noteTitleWithOpenAction = `<span class="noteTitle"><a href="${noteOpenActionURL}"><i class="fa-regular fa-file-lines pad-right"></i> ${folderNamePart}${titlePart}</a></span>`
-        // TODO: if possible change to use internal links: see method in Dashboard
-        // see discussion at https://discord.com/channels/763107030223290449/1007295214102269982/1016443125302034452
-        // const noteTitleWithOpenAction = `<button onclick=openNote()>${folderNamePart}${titlePart}</button>`
-
-        if (this.isCompleted) {
-          return `<span class="checked">${noteTitleWithOpenAction}</span>`
-        } else if (this.isCancelled) {
-          return `<span class="cancelled">${noteTitleWithOpenAction}</span>`
-        } else if (this.isPaused) {
-          return `<span class="paused">Paused: ${noteTitleWithOpenAction}</span>`
-        } else {
-          return `${noteTitleWithOpenAction}`
-        }
-      }
-
-      case 'Markdown': {
-        if (this.isCompleted) {
-          return `[x] ${folderNamePart}[[${titlePart}]]`
-        } else if (this.isCancelled) {
-          return `[-] ${folderNamePart}[[${titlePart}]]`
-        } else if (this.isPaused) {
-          return `⏸ **Paused**: ${folderNamePart}[[${titlePart}]]`
-        } else {
-          return `${folderNamePart}[[${titlePart}]]` // if this has a [ ] prefix then it of course turns it into a task, which is probably not what we want.
-        }
-      }
-
-      case 'list': {
-        if (this.isCompleted) {
-          return `${folderNamePart}[[${titlePart}]]`
-        } else if (this.isCancelled) {
-          return `~~${folderNamePart}[[${titlePart}]]~~`
-        } else if (this.isPaused) {
-          return `⏸ **Paused**: ${folderNamePart}[[${titlePart}]]`
-        } else {
-          return `${folderNamePart}[[${titlePart}]]` // if this has a [ ] prefix then it of course turns it into a task, which is probably not what we want.
-        }
-      }
-
-      default:
-        logWarn('Project::decoratedProjectTitle', `Unknown style '${style}'; nothing returned.`)
-        return ''
-    }
-  }
-
-  /**
-   * Returns line showing more detailed summary of the project, for output in Rich (HTML) or Markdown formats or simple list format.
-   * Now uses fontawesome icons for some indicators.
-   * @param {string} style
-   * @param {boolean} includeFolderName
-   * @param {boolean?} displayDates
-   * @param {boolean?} displayProgress
-   * @returns {string}
-   */
-  detailedSummaryLine(style: string, includeFolderName: boolean, displayDates: boolean = true, displayProgress: boolean = true): string {
-    let output = ''
-    const thisPercent = (isNaN(this.percentComplete)) ? '0%' : ` ${this.percentComplete}%`
-    const totalTasksStr = (this.completedTasks + this.openTasks).toLocaleString()
-    const statsProgress = `${thisPercent} done (of ${totalTasksStr} ${(this.completedTasks + this.openTasks > 1) ? 'tasks' : 'task'})`
-
-    if (style === 'Rich') {
-      output = '\t<tr class="projectRow">\n\t\t'
-
-      // Column 1: circle indicator
-      if (this.isCompleted) {
-        output += `<td class="first-col-indicator checked">${this.addFAIcon('fa-solid fa-circle-check')}</td>` // ('checked' gives colour)
-      }
-      else if (this.isCancelled) {
-        output += `<td class="first-col-indicator cancelled">${this.addFAIcon('fa-solid fa-circle-xmark')}</td>` // ('cancelled' gives colour)
-      }
-      else if (this.isPaused) {
-        output += `<td class="first-col-indicator">${this.addFAIcon("fa-solid fa-circle-pause", "#888888")}</td>`
-      }
-      else if (isNaN(this.percentComplete)) {
-        output += `<td class="first-col-indicator">${this.addFAIcon('fa-solid fa-circle', '#888888')}</td>`
-      }
-      else if (this.percentComplete === 0) {
-        output += `<td class="first-col-indicator">${this.addSVGPercentRing(100, '#FF000088', '0')}</td>`
-      }
-      else {
-        output += `<td class="first-col-indicator">${this.addSVGPercentRing(this.percentComplete, 'multicol', String(this.percentComplete))}</td>`
-      }
-
-      // Column 2a: Project name / link / edit dialog trigger button
-      const editButton = `          <a class="dialogTrigger" onclick="showProjectControlDialog({encodedFilename: '${encodeRFC3986URIComponent(this.filename)}'})"><i class="fa-light fa-edit pad-left"></i></a>\n`
-      if (this.isCompleted || this.isCancelled || this.isPaused) {
-        output += `<td>${this.decoratedProjectTitle(style, includeFolderName)}&nbsp;${editButton}`
-      }
-      else if (this.percentComplete === 0 || isNaN(this.percentComplete)) {
-        output += `<td>${this.decoratedProjectTitle(style, includeFolderName)}&nbsp;${editButton}`
-      } else {
-        output += `\n\t\t\t<td>${this.decoratedProjectTitle(style, includeFolderName)}&nbsp;${editButton}`
-      }
-
-      // Column 2b: progress information
-      if (displayProgress && !this.isCompleted && !this.isCancelled) {
-        // logDebug('Project::detailedSummaryLine', `'${this.lastProgressComment}' / ${statsProgress} for ${this.title}`)
-        // Add this.lastProgressComment (if it exists) on line under title (and project is still open)
-        if (displayDates) {
-          if (this.lastProgressComment !== '') {
-            output = `${output}<br /><i class="fa-regular fa-info-circle fa-sm pad-right"></i> ${this.lastProgressComment}</td>`
-          } else {
-            output = `${output}<br />${statsProgress}</td>`
-          }
-        } else {
-          // write progress in next cell instead
-          if (this.lastProgressComment !== '') {
-            output += `</td>\n\t\t\t<td><i class="fa-regular fa-info-circle fa-sm pad-right"></i> ${this.lastProgressComment}</td>`
-          } else {
-            output += `</td>\n\t\t\t<td>${statsProgress}</td>`
-          }
-        }
-      }
-
-      // Columns 3/4: date information
-      if (displayDates && !this.isPaused) {
-        if (this.isCompleted) {
-          // "completed after X"
-          const completionRef = (this.completedDuration)
-            ? this.completedDuration
-            : "completed"
-          output += `<td colspan=2 class="checked">Completed ${completionRef}</td>`
-        } else if (this.isCancelled) {
-          // "cancelled X ago"
-          const cancellationRef = (this.cancelledDuration)
-            ? this.cancelledDuration
-            : "cancelled"
-          output += `<td colspan=2 class="cancelled">Cancelled ${cancellationRef}</td>`
-        }
-        if (!this.isCompleted && !this.isCancelled) {
-          output = (this.nextReviewDays != null && !isNaN(this.nextReviewDays))
-            ? (this.nextReviewDays > 0)
-              ? `${output}<td>${localeRelativeDateFromNumber(this.nextReviewDays)}</td>`
-              : `${output}<td><p><b>${localeRelativeDateFromNumber(this.nextReviewDays)}</b></p></td>` // the <p>...</p> is needed to trigger bold colouring (if set)
-            : `${output}<td></td>`
-          output = (this.dueDays != null && !isNaN(this.dueDays))
-            ? (this.dueDays > 0)
-              ? `${output}<td>${localeRelativeDateFromNumber(this.dueDays)}</td>`
-              : `${output}<td><p><b>${localeRelativeDateFromNumber(this.dueDays)}</b></p></td>` // the <p>...</p> is needed to trigger bold colouring (if set)
-            : `${output}<td></td>`
-        }
-      } else {
-        output += '<td></td><td></td>' // to avoid layout inconsistencies
-      }
-      output += '\n\t</tr>'
-    }
-    else if (style === 'Markdown' || style === 'list') {
-      output = '- '
-      output += `${this.decoratedProjectTitle(style, includeFolderName)}`
-      // logDebug('', `${this.decoratedProjectTitle(style, includeFolderName)}`)
-      if (displayDates && !this.isPaused) {
-        if (this.isCompleted) {
-          // completed after X or cancelled X ago, depending
-          const completionRef = (this.completedDuration)
-            ? this.completedDuration
-            : "completed"
-          output += `\t(Completed ${completionRef})`
-        } else if (this.isCancelled) {
-          // completed after X or cancelled X ago, depending
-          const cancellationRef = (this.cancelledDuration)
-            ? this.cancelledDuration
-            : "cancelled"
-          output += `\t(Cancelled ${cancellationRef})`
-        }
-      }
-      if (displayProgress && !this.isCompleted && !this.isCancelled) {
-        // Show progress comment if available ...
-        if (this.lastProgressComment !== '' && !this.isCompleted && !this.isCancelled) {
-          output += `\t${thisPercent} done: ${this.lastProgressComment}`
-        }
-        // ... else show stats
-        else {
-          output += `\t${statsProgress}`
-        }
-      }
-      if (displayDates && !this.isPaused && !this.isCompleted && !this.isCancelled) {
-        output = (this.dueDays != null && !isNaN(this.dueDays)) ? `${output}\tdue ${localeRelativeDateFromNumber(this.dueDays)}` : output
-        output =
-          (this.nextReviewDays != null && !isNaN(this.nextReviewDays))
-            ? this.nextReviewDays > 0
-              ? `${output}\tReview ${localeRelativeDateFromNumber(this.nextReviewDays)}`
-              : `${output}\tReview due **${localeRelativeDateFromNumber(this.nextReviewDays)}**`
-            : output
-      }
-    } else {
-      logWarn('Project::detailedSummaryLine', `Unknown style '${style}'; nothing returned.`)
-      output = ''
-    }
-    return output
-  }
-
-  /**
-   * Add SVG ready for percent ring with the number in the middle.
-   * Note: this is kept in this file as it is specific to Review functionality. But it relies on the more generic 'makeSVGPercentRing' helper function.
-   * Note: It needs to be followed by call to JS function setPercentRing() to set the ring's state.
-   * @param {number} percent 0-100
-   * @param {string?} color for ring and text (as colour name or #RGB), or 'multicol' to mean shading between red and green
-   * @param {string?} textToShow inside ring, which can be different from just the percent, which is used by default
-   * @returns {string} SVG code to insert in HTML
-   */
-  addSVGPercentRing(percent: number, colorIn: string = 'multicol', text: string = ''): string {
-    const textToShow = (text !== '') ? text : String(percent)
-    const colorToUse = (colorIn === 'multicol')
-      ? redToGreenInterpolation(percent)
-      : colorIn
-    return makeSVGPercentRing(percent, colorToUse, textToShow, this.ID)
-  }
-
-  /**
-   * Insert a fontawesome icon in given color.
-   * Other styling comes from CSS for 'circle-icon' (just sets size)
-   * Note: it doesn't put item in a filled circle; just so far I've picked icons that look like that.
-   * @param {string} faClasses CSS class name(s) to use for FA icons
-   * @param {string} colorStr optional
-   * @returns HTML string to insert
-   */
-  addFAIcon(faClasses: string, colorStr: string = ''): string {
-    if (colorStr !== '') {
-      return `<span class="${faClasses} circle-icon" style="color: ${colorStr}"></span>`
-    } else {
-      return `<span class="${faClasses} circle-icon"></span>`
-    }
-  }
-}
-
-/**
- * Form HTML for a 'fake' button that is used to call (via x-callback) one of this plugin's commands.
- * Note: this is not a real button, bcause at the time I started this real <button> wouldn't work in NP HTML views, and Eduard didn't know why.
- * @param {string} buttonText to display on button
- * @param {string} commandName to call when button is 'clicked'
- * @param {string?} tooltipText to hover display next to button
- * @returns {string}
- */
-export function makeFakeButton(buttonText: string, commandName: string, commandArgs: string, tooltipText: string = ''): string {
-  const xcallbackURL = createRunPluginCallbackUrl('jgclark.Reviews', commandName, commandArgs)
-  const output = (tooltipText)
-    ? `<span class="fake-button tooltip"><a class="button" href="${xcallbackURL}">${buttonText}</a><span class="tooltiptext">${tooltipText}</span></span>`
-    : `<span class="fake-button"><a class="button" href="${xcallbackURL}">${buttonText}</span>`
-  return output
-}
-
 /**
  * WARNING: DEPRECATED
  * Function to save changes to the Editor to the cache to be available elsewhere straight away.
- * Note: From 3.9.3 there's a function for this, but we needed something else before then (now removed) that did a basic 1s wait.
+ * Note: Now declared v3.9.3 as minimum version, so we can use API function for this.
  */
 // eslint-disable-next-line no-unused-vars
 export async function saveEditorToCache(completed: any): Promise<void> {
-  try {
-    // // If 3.9.3alpha or later call specific new function
-    // if (NotePlan.environment.buildVersion > 1049) {
-    logDebug('saveEditorToCache', 'waiting for Editor.save ...')
-    await Editor.save()
-    logDebug('saveEditorToCache', '... done')
-    // }
-    // // else wait for 1 second
-    // else {
-    //   logDebug('saveEditorToCache', 'waiting for 1 second ...')
-    //   setTimeout(() => {
-    //     DataStore.updateCache(Editor.note, true)
-    //     completed()
-    //   }, 1000)
-    //   logDebug('saveEditorToCache', '... done')
-    // }
-  } catch (error) {
-    logError('saveEditorToCache', error.message)
-  }
+  logDebug('saveEditorToCache', 'waiting for Editor.save ...')
+  await Editor.save()
+  logDebug('saveEditorToCache', '... done')
 }
-
-// Looks like this doesn't work in NP. See some Discord chat @EduardMe, 19.6.2023
-// Error message "NotePlan_Beta.JSPromiseConstructor is not a constructor (evaluating 'new Promise((resolve => setTimeout(resolve, milliseconds)))')"
-// function delay(milliseconds: number) {
-//   try {
-//     // $FlowIgnore - @EduardMe says NP overrides the Promise mechanism
-//     const promise = new Promise()
-//     setTimeout(() => { promise.success() }, milliseconds)
-//     return promise
-//   } catch (error) {
-//     logError('delay', error.message)
-//   }
-// }
-
-// Note: commented out, as not currently used, and I can't remember what this was for
-// const delay = (ms: number) => {
-//   const start = Date.now()
-//   let now = start
-//   while (now - start < ms) {
-//     now = Date.now()
-//   }
-// }
-
 
 //-------------------------------------------------------------------------------
 /**
  * Update project metadata @mentions (e.g. @reviewed(date)) in the metadata line of the note in the Editor.
- * It takes each mention in the array (e.g. '@reviewed(2023-06-23)') and all other versions of @reviewed will be removed first, before that string is appended.
+ * It takes each mention in the array (e.g. '@reviewed(2023-06-23)') and all other versions of it will be removed first, before that string is appended.
  * @author @jgclark
  * @param {Array<string>} mentions to update:
  * @returns { ?TNote } current note
@@ -1345,6 +462,8 @@ export function updateMetadataInNote(note: TNote, updatedMetadataArr: Array<stri
 }
 
 //-------------------------------------------------------------------------------
+// Other helpers
+
 /**
  * Update project metadata @mentions (e.g. @reviewed(date)) in the note in the Editor
  * @author @jgclark
@@ -1441,6 +560,21 @@ export function deleteMetadataMentionInNote(noteToUse: TNote, mentionsToDeleteAr
 export function updateDashboardIfOpen(): void {
   // Finally, refresh Dashboard. Note: Designed to fail silently if it isn't installed, or open.
   const refreshXCallbackURL = createRunPluginCallbackUrl('jgclark.Dashboard', 'refreshProjectSection', '')
-  logDebug('makeFullReviewList', `sent message to refresh 🎛 Dashboard: ${refreshXCallbackURL}`)
+  logDebug('updateDashboardIfOpen', `sent message to refresh 🎛 Dashboard: ${refreshXCallbackURL}`)
   NotePlan.openURL(refreshXCallbackURL) // no point in having await
+}
+
+/**
+ * Insert a fontawesome icon in given color.
+ * Other styling comes from CSS for 'circle-icon' (just sets size)
+ * @param {string} faClasses CSS class name(s) to use for FA icons
+ * @param {string} colorStr optional
+ * @returns HTML string to insert
+ */
+export function addFAIcon(faClasses: string, colorStr: string = ''): string {
+  if (colorStr !== '') {
+    return `<span class="${faClasses}" style="color: ${colorStr}"></span>`
+  } else {
+    return `<span class="${faClasses}"></span>`
+  }
 }
